size unpacking fix
Sizes were being read as doubles; they need to be read as longs w/ four
pad bits. alen/ssize/initsize have all been checked as working;
start_vcn and last_vcn have not been checked but should theoretically
follow similar convention.

There is a chance that the sizes should be read in as signed longs or
ints. There didn't seem to be a difference when I tested it, but if
strange values appear, this could be the issue.